pyproject.toml
[project]
name = "aurayouth"
version = "0.1.0"
description = "AI-Powered Mental Wellness Platform for Youth"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.12",
    "opencv-python>=4.10.0",
    "librosa>=0.10.2",
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    "pymongo>=4.10.0",
    "motor>=3.6.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0,<4.1.0",
    "passlib[bcrypt]>=1.7.4,<1.8.0",
    "python-dotenv>=1.0.0",
]
